# helper/api.yaml
openapi: 3.0.0

info:
  title: GGGGGGGGGG API
  version: 1.0.0

servers:
  - url: http://localhost:8080/api

tags:
  - name: GGGGGGGGGG

paths:
  /lllllllllls:
    get:
      tags:
        - GGGGGGGGGG
      summary: Get all GGGGGGGGGG
      operationId: getAllGGGGGGGGGG
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: List of GGGGGGGGGG
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GGGGGGGGGGDTO'
    post:
      tags:
        - GGGGGGGGGG
      summary: Create a new GGGGGGGGGG
      operationId: createGGGGGGGGGG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GGGGGGGGGGDTO'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GGGGGGGGGGDTO'

  /lllllllllls/{id}:
    get:
      tags:
        - GGGGGGGGGG
      summary: Get GGGGGGGGGG by ID
      operationId: getGGGGGGGGGGById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GGGGGGGGGGDTO'
    put:
      tags:
        - GGGGGGGGGG
      summary: Update GGGGGGGGGG by ID
      operationId: updateGGGGGGGGGGById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GGGGGGGGGGDTO'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GGGGGGGGGGDTO'
    delete:
      tags:
        - GGGGGGGGGG
      summary: Delete GGGGGGGGGG by ID
      operationId: deleteGGGGGGGGGGById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GGGGGGGGGGDTO'

components:
  schemas:
    GGGGGGGGGGDTO:
      type: object
      properties: